Moved Clean to submodule
Since the clean function will also depend on the engine, moved the logic
of this to the `tex` submodule.
